+++ test/src/unit-cppapi-schema.cc
@@ -177,20 +177,11 @@ TEST_CASE("C++ API: Schema", "[cppapi]") {
 
   SECTION("Invalid dimension types") {
     Domain dom(ctx);
-    dom.add_dimension(Dimension::create<int>(ctx, "d1", {{0, 10}}));
+    dom.add_dimension(Dimension::create<int>(ctx, "d1", {{0, 10}}, 11));
     CHECK_THROWS_AS(
-        dom.add_dimension(Dimension::create<uint64_t>(ctx, "d2", {{0, 10}})),
+        dom.add_dimension(
+            Dimension::create<uint64_t>(ctx, "d2", {{0, 10}}, 11)),
         TileDBError);
-  }
-
-  SECTION("Dimensions without tile extent") {
-    Domain dom(ctx);
-    dom.add_dimension(Dimension::create<int>(ctx, "d1", {{0, 10}}));
-    ArraySchema schema(ctx, TILEDB_DENSE);
-    schema.set_domain(dom);
-
-    auto dom_get = schema.domain();
-    CHECK(dom_get.dimensions()[0].tile_extent<int>() == 11);
   }
 }