+++ tiledb/sm/array_schema/dimension.cc
@@ -213,6 +213,13 @@ Status Dimension::set_tile_extent(const void* tile_extent) {
     return Status::DimensionError(
         "Cannot set tile extent; Domain must be set first");
 
+  // Note: this check was added in release 1.6.0. Older arrays may have been
+  // serialized with a null extent, and so it is still supported internally.
+  // But users can not construct dimension objects with null tile extents.
+  if (tile_extent == nullptr)
+    return LOG_STATUS(Status::DimensionError(
+        "Cannot set tile extent; tile extent cannot be null"));
+
   std::free(tile_extent_);
   if (tile_extent == nullptr) {
     tile_extent_ = nullptr;
@@ -436,7 +443,7 @@ Status Dimension::check_tile_extent() const {
 
   // Check if tile extent exceeds domain
   if (!is_int) {
-    if (*tile_extent > domain[1] - domain[0])
+    if (*tile_extent > (domain[1] - domain[0] + 1))
       return LOG_STATUS(
           Status::DimensionError("Tile extent check failed; Tile extent "
                                  "exceeds dimension domain range"));