+++ tiledb/sm/cpp_api/dimension.h
@@ -346,38 +346,6 @@ class Dimension {
   }
 
   /**
-   * Factory function for creating a new dimension with datatype T with no
-   * tile extent. TileDB uses a tile extent equal to the whole domain extent in
-   * this dimension.
-   *
-   * **Example:**
-   * @code{.cpp}
-   * tiledb::Context ctx;
-   * // Create a dimension with inclusive domain [0,1000] and implicit
-   * // tile extent 1001.
-   * auto dim = Dimension::create<int32_t>(ctx, "d", {{0, 1000}});
-   * @endcode
-   *
-   * @tparam T int, char, etc...
-   * @param ctx The TileDB context.
-   * @param name The dimension name.
-   * @param domain The dimension domain. A pair [lower,upper] of inclusive
-   * bounds.
-   * @return A new `Dimension` object.
-   */
-  template <typename T>
-  static Dimension create(
-      const Context& ctx,
-      const std::string& name,
-      const std::array<T, 2>& domain) {
-    using DataT = impl::TypeHandler<T>;
-    static_assert(
-        DataT::tiledb_num == 1,
-        "Dimension types cannot be compound, use arithmetic type.");
-    return create_impl(ctx, name, DataT::tiledb_type, &domain, nullptr);
-  }
-
-  /**
    * Factory function for creating a new dimension (non typechecked).
    *
    * @param ctx The TileDB context.